python/agents/data_science_powerpoint/pyproject.toml
[tool.poetry]
name = "data-science-powerpoint"
version = "0.1"
description = "Data Science and Data QnA Multi-Agent"
authors = ["Meltem Subasioglu <msubasioglu@google.com>"]
license = "Apache License 2.0"
readme = "README.md"
package-mode = true


[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
google-adk = "^1.5.0"
immutabledict = "^4.2.1"
sqlglot = "^26.10.1"
db-dtypes = "^1.4.2"
regex = "^2024.11.6"
tabulate = "^0.9.0"
fastapi = "^0.115.0" # Explicitly adding fastapi to try and resolve conflict
httpx = "^0.27.0" # For making HTTP requests to the RAG service
python-pptx = "^0.6.21"
google-cloud-storage = "^2.10.0"
google-cloud-aiplatform = { extras = [
    "adk",
    "agent-engines",
    "prediction", # For embedding models
    "matching-engine", # For vector search
], version = "^1.93.0" }
absl-py = "^2.2.2"
pydantic = "^2.11.3"
pandas = "^2.3.0"
numpy = "^2.3.1"

[tool.poetry.group.dev.dependencies]
google-cloud-aiplatform = { extras = [
    "adk",
    "agent-engines",
    "evaluation",
], version = "^1.93.0" }
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
google-adk = { extras = ["eval"], version = "^1.5.0" }


[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
#addopts = "-vv -s --pdb"
testpaths = ["tests/"]
log_level = "DEBUG"
log_cli = true
log_auto_indent = true
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_cli_format = "[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    #"error",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    #'ignore:function ham\(\) is deprecated:DeprecationWarning',
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
